# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TechniConnect - Electronics Technician & Shop Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories
      parameters:
        - name: type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  # SHOPS
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: shopId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # SERVICES
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List technician services
      parameters:
        - name: technicianId
          in: query
          schema:
            type: integer
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    put:
      operationId: updateService
      tags: [services]
      summary: Update service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  # BOOKINGS
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      summary: List reviews for a target
      parameters:
        - name: targetType
          in: query
          schema:
            type: string
        - name: targetId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"
    post:
      operationId: createReview
      tags: [reviews]
      summary: Create review
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"

  # ADMIN
...
  /location/nearby-shops:
    get:
      operationId: getNearbyShops
      tags: [location]
      summary: Get nearby shops within radius
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Nearby shops
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ShopWithDistance"
  /location/nearby-technicians:
    get:
      operationId: getNearbyTechnicians
      tags: [location]
      summary: Get nearby technicians within radius
      parameters:
        - name: lat
          in: query
          required: true
          schema:
            type: number
        - name: lng
          in: query
          required: true
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Nearby technicians
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TechnicianWithDistance"
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        type:
          type: string
      required: [id, name, type]
    CategoryInput:
      type: object
      properties:
        name:
          type: string
        icon:
          type: string
        type:
          type: string
          enum: [shop, technician, product]
      required: [name, type]
    Shop:
      type: object
      properties:
        id:
          type: integer
        ownerId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        rating:
          type: number
        ratingCount:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        workingHours:
          type: ["string", "null"]
        createdAt:
          type: string
        ownerName:
          type: ["string", "null"]
      required: [id, ownerId, name, address, lat, lng, rating, ratingCount, isVerified, isActive, createdAt]
...
    ShopWithDistance:
      allOf:
        - $ref: "#/components/schemas/Shop"
        - type: object
          properties:
            distance:
              type: number
          required: [distance]
...
    Technician:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        bio:
          type: ["string", "null"]
        skills:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
        hourlyRate:
          type: number
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
        isAvailable:
          type: boolean
        rating:
          type: number
        ratingCount:
          type: integer
        isVerified:
          type: boolean
        isActive:
          type: boolean
        serviceRadius:
          type: integer
        profileImage:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, name, skills, experienceYears, hourlyRate, isAvailable, rating, ratingCount, isVerified, isActive, serviceRadius, createdAt]
...
    TechnicianWithDistance:
      allOf:
        - $ref: "#/components/schemas/Technician"
        - type: object
          properties:
            distance:
              type: number
          required: [distance]
...
    Product:
      type: object
      properties:
        id:
          type: integer
        shopId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: ["string", "null"]
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
        shopName:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, shopId, name, price, stock, isActive, createdAt]
    ProductInput:
      type: object
      properties:
        shopId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
      required: [shopId, name, price, stock]
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        images:
          type: array
          items:
            type: string
        stock:
          type: integer
        isActive:
          type: boolean
    ProductListResponse:
      type: object
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [products, total, page, limit]
    Service:
      type: object
      properties:
        id:
          type: integer
        technicianId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        category:
          type: ["string", "null"]
        durationMinutes:
          type: integer
        isActive:
          type: boolean
        createdAt:
          type: string
      required: [id, technicianId, name, price, durationMinutes, isActive, createdAt]
    ServiceInput:
      type: object
      properties:
        technicianId:
          type: integer
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        durationMinutes:
          type: integer
      required: [technicianId, name, price, durationMinutes]
    ServiceUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        durationMinutes:
          type: integer
        isActive:
          type: boolean
...
    Review:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        targetId:
          type: integer
        targetType:
          type: string
          enum: [technician, shop]
        rating:
          type: integer
        comment:
          type: ["string", "null"]
        userName:
          type: ["string", "null"]
        userImage:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, userId, targetId, targetType, rating, createdAt]
    ReviewInput:
      type: object
      properties:
        targetId:
          type: integer
        targetType:
          type: string
          enum: [technician, shop]
        rating:
          type: integer
        comment:
          type: string
      required: [targetId, targetType, rating]
...