# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechniConnect - Electronics Technician & Shop Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  shopId: number;
  name: string;
  description?: string;
  price: number;
  category?: string;
  images?: string[];
  stock: number;
}