# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TechniConnect - Electronics Technician & Shop Marketplace API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  technicianId: number;
  name: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  category?: string | null;
  durationMinutes: number;
  isActive: boolean;
  createdAt: string;
}